pyproject.toml
[tool.poetry]
name = "interactive-brokers-products-database"
version = "0.1.0"
description = ""
authors = ["Christophe <christophe.alexandre@pm.me>"]
readme = "README.md"

# imported from https://cloud.google.com/functions/docs/writing/specifying-dependencies-python#python38
[tool.poetry.dependencies]
python = "^3.11"
click = "8.0.3"
cloudevents = "1.2.0"
deprecation = "2.1.0"
gunicorn = "20.1.0"
itsdangerous = "2.0.1"
Jinja2 = "3.0.3"
MarkupSafe = "2.0.1"
packaging = "21.3"
pathtools = "0.1.2"
pyparsing= "3.0.6"
setuptools = "60.3.1"
watchdog = "1.0.2"
Werkzeug = "2.0.2"
google-auth = "2.17.3"
google-cloud-datastore = "2.15.1"
google-cloud-logging = "^3.6.0"
functions-framework = "3.0.0"
google-cloud-tasks = "^2.14.1"

[tool.poetry.group.dev.dependencies]
Flask = "2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
